+++ services/src/main/java/org/keycloak/protocol/oidc/tokenexchange/FederatedTokenExchangeProviderFactory.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2021 Red Hat, Inc. and/or its affiliates
+ * Copyright 2025 Red Hat, Inc. and/or its affiliates
  *  and other contributors as indicated by the @author tags.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
@@ -11,43 +11,61 @@
  *  Unless required by applicable law or agreed to in writing, software
  *  distributed under the License is distributed on an "AS IS" BASIS,
  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  *
  */
-package org.keycloak.protocol.oidc;
+
+package org.keycloak.protocol.oidc.tokenexchange;
 
 import org.keycloak.Config;
+import org.keycloak.common.Profile;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.KeycloakSessionFactory;
+import org.keycloak.protocol.oidc.TokenExchangeProvider;
+import org.keycloak.protocol.oidc.TokenExchangeProviderFactory;
+import org.keycloak.provider.EnvironmentDependentProviderFactory;
 
 /**
- * Default token exchange provider factory
+ * Provider factory for external-internal or internal-external token exchange, where identity providers are involved
  *
- * @author <a href="mailto:dmitryt@backbase.com">Dmitry Telegin</a>
+ * @author <a href="mailto:mposolda@redhat.com">Marek Posolda</a>
  */
-public class DefaultTokenExchangeProviderFactory implements TokenExchangeProviderFactory {
+public class FederatedTokenExchangeProviderFactory implements TokenExchangeProviderFactory, EnvironmentDependentProviderFactory {
 
     @Override
     public TokenExchangeProvider create(KeycloakSession session) {
-        return new DefaultTokenExchangeProvider();
+        return new FederatedTokenExchangeProvider();
     }
 
     @Override
     public void init(Config.Scope config) {
+
     }
 
     @Override
     public void postInit(KeycloakSessionFactory factory) {
+
     }
 
     @Override
     public void close() {
+
     }
 
     @Override
     public String getId() {
-        return "default";
+        return "federated";
     }
 
+    @Override
+    public boolean isSupported(Config.Scope config) {
+        return Profile.isFeatureEnabled(Profile.Feature.TOKEN_EXCHANGE_FEDERATED_V2);
+    }
+
+    @Override
+    public int order() {
+        return 2;
+    }
 }
